+++ src/generator/angular-file-type.model.ts
@@ -1,0 +1,11 @@
+export const NG_FILE_TYPE = {
+  module: 'module',
+  service: 'service',
+  component: 'component',
+} as const;
+
+export const NG_FILE_TYPES: NgFileType[] = Object.keys(
+  NG_FILE_TYPE
+) as NgFileType[];
+
+export type NgFileType = keyof typeof NG_FILE_TYPE;