+++ CHANGELOG.md
@@ -2,6 +2,17 @@
 
 ## [Unreleased]
 
+## [0.0.2]
+
+### Changed
+
+- Read angular.json to grab default selector name
+- Move menu item from the very top of the list into it own group instead
+
+### Fixed
+
+- Generated files will not replace existing ones
+
 ## [0.0.1]
 
 - Initial release